% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{data}
\name{hsdependentdays}
\alias{hsdependentdays}
\title{Making the parameter data frame for dependents (pups)}
\format{Data frames with columns
\describe{
\item{hsjuvedays}{3 column data frame containing parameters governing the survival curve for dependents: 1st is effect on survival, 2nd is days tolerated, 3rd is days before max effect}
}}
\source{
Interim PCOD Report
}
\usage{
hsdependentdays
}
\description{
Making the parameter data frame for dependents (pups)
}
\examples{
hsdependentdays
}
\keyword{datasets}

