% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{data}
\name{hsmaturedays}
\alias{hsmaturedays}
\title{Getting the elicitation data prepared into an R package}
\format{Data frames with columns
\describe{
\item{hsmaturedays}{3 column data frame containing parameters governing the survival curve: 1st is effect on survival, 2nd is days tolerated, 3rd is days before max effect}
}}
\source{
Interim PCOD Report
}
\usage{
hsmaturedays
}
\description{
Getting the elicitation data prepared into an R package
}
\examples{
hsmaturedays
}
\keyword{datasets}

