% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{data}
\name{hsoutput_dependent}
\alias{hsoutput_dependent}
\title{Making the parameter data frame for dependents (pups)}
\format{Data frames with columns
\describe{
\item{hsoutput_dependent}{2 element list containging:}
}}
\source{
Interim PCOD Report
}
\usage{
hsoutput_dependent
}
\description{
Making the parameter data frame for dependents (pups)
}
\examples{
hsoutput_dependent
}
\keyword{datasets}

