% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{data}
\name{hsoutput_juvenile}
\alias{hsoutput_juvenile}
\title{Making the parameter data frame for Juveniles}
\format{Data frames with columns
\describe{
\item{hsoutput_juvenile}{2 element list containing:}
}}
\source{
Interim PCOD Report
}
\usage{
hsoutput_juvenile
}
\description{
Making the parameter data frame for Juveniles
}
\examples{
hsoutput_juvenile
}
\keyword{datasets}

