% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{data}
\name{hsmaturedays}
\alias{hsmaturedays}
\title{Getting the elicitation data prepared into an R package}
\format{Data frames with columns
\describe{
\item{hsmaturedays}{3 Parameters governing the survival curve}
}}
\source{
Interim PCOD Report
}
\usage{
hsmaturedays
}
\description{
Getting the elicitation data prepared into an R package
}
\examples{
hsmaturedays
}
\keyword{datasets}

